dev_0.1.0/requirements.txt
contourpy==1.1.0
cycler==0.11.0
filterpy==1.4.5
fonttools==4.41.1
importlib-resources==6.0.0
joblib==1.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
RPi.GPIO==0.7.1
scikit-learn==1.2.2
scipy==1.11.1
six==1.16.0
smbus2==0.4.2
threadpoolctl==3.2.0
tzdata==2023.3
zipp==3.16.2
